Arrgelo de:
	Sistema de eleccion de color y talla de articulo ya funcional
	Cambio en el campo contraseña por contraseya, debido a posibles fallos
	Cambio de los JARS de rutas absolutas a relativas para evitar fallos de
imports
	Arreglado metodo de eliminarArticulo de la cesta y cambio de precio
total de la venta

+++ DeustoWear/src/clases/Camiseta.java
@@ -24,7 +24,7 @@ public class Camiseta extends Articulo{
 	 * @param sexo
 	 * @param imagen
 	 */
-	public Camiseta(int id, String name,String talla, int precio, String color,String sexo, String imagen) {
+	public Camiseta(int id, String name,String talla, double precio, String color,String sexo, String imagen) {
 		super(id,name,talla,precio,color,sexo,imagen);
 		
 	}
